# services/mapi-xdg/docs/indexer/swagger.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Dogecoin - Blockchain Indexer API",
    "description": "Dogecoin indexer endpoints: addresses, dunes, inscriptions and DRC-20. The same binary also proxies a set of node RPC endpoints (blocks, mempool, chain info) documented in the repository README.",
    "license": {
      "name": "Apache 2.0",
      "url": "https://www.apache.org/licenses/LICENSE-2.0.txt"
    },
    "version": "0.0.1"
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Local instance"
    }
  ],
  "paths": {
...
    "/addresses/{address}/utxos": {
      "get": {
        "tags": [
          "Addresses"
        ],
        "summary": "UTxOs by Address",
        "description": "List of all UTxOs which reside at the specified address or script pubkey.",
        "operationId": "utxos_by_address",
        "parameters": [
          {
            "name": "address",
            "in": "path",
            "description": "Dogecoin address or hex encoded script pubkey",
            "required": true,
            "schema": {
              "type": "string"
            },
            "example": "DNG3G7pKc1DgciyWs36GFUnVyvJieDhKdb"
          },
          {
            "name": "filter_dust",
            "in": "query",
            "description": "Ignore UTxOs containing less than 100000 shibes",
            "required": false,
            "schema": {
              "type": "boolean",
              "nullable": true
            }
          },
          {
            "name": "filter_dust_threshold",
            "in": "query",
            "description": "Ignore UTxOs containing less than specified shibes",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int64",
              "nullable": true,
              "minimum": 0
            }
          },
          {
            "name": "exclude_metaprotocols",
            "in": "query",
            "description": "Exclude metaprotocol UTxOs (UTxOs involved in metaprotocols like dunes or inscriptions)",
            "required": false,
            "schema": {
              "type": "boolean",
              "nullable": true
            }
          },
          {
            "name": "count",
            "in": "query",
            "description": "The max number of results per page",
            "required": false,
            "schema": {
              "allOf": [
                {
                  "type": "integer",
                  "default": 100,
                  "minimum": 0
                }
              ],
              "nullable": true
            }
          },
          {
            "name": "order",
            "in": "query",
            "description": "The order in which the results are sorted (by height at which UTxO was produced)",
            "required": false,
            "schema": {
              "allOf": [
                {
                  "type": "string",
                  "default": "asc",
                  "enum": [
                    "asc",
                    "desc"
                  ]
                }
              ],
              "nullable": true
            }
          },
          {
            "name": "from",
            "in": "query",
            "description": "Return only UTxOs created on or after a specific height",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int64",
              "nullable": true,
              "minimum": 0
            }
          },
          {
            "name": "to",
            "in": "query",
            "description": "Return only UTxOs created on or before a specific height",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int64",
              "nullable": true,
              "minimum": 0
            }
          },
          {
            "name": "cursor",
            "in": "query",
            "description": "Pagination cursor string, use the cursor included in a page of results to fetch the next page",
            "required": false,
            "schema": {
              "type": "string",
              "nullable": true
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Requested data",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaginatedUtxo"
                },
                "example": {
                  "data": [
                    {
                      "address": "DEvbE2KyPKAzDsj511wijyF1NSBN9qMcix",
                      "confirmations": 121,
                      "dunes": [
                        {
                          "amount": "5.00000000",
                          "dune_id": "5095478:8"
                        }
                      ],
                      "height": 5095478,
                      "inscriptions": [],
                      "satoshis": "100000",
                      "script_pubkey": "76a9146b53cbbe67beef7079e6ea3e39bdfdc34bf2707b88ac",
                      "txid": "ee4b62b6e221eb92ea821ecb7932eb325647d9fdce5632747905146a561b8a4b",
                      "vout": 1
                    },
                    {
                      "address": "DEvbE2KyPKAzDsj511wijyF1NSBN9qMcix",
                      "confirmations": 121,
                      "dunes": [],
                      "height": 5095478,
                      "inscriptions": [],
                      "satoshis": "1999441400000",
                      "script_pubkey": "76a9146b53cbbe67beef7079e6ea3e39bdfdc34bf2707b88ac",
                      "txid": "ee4b62b6e221eb92ea821ecb7932eb325647d9fdce5632747905146a561b8a4b",
                      "vout": 2
                    }
                  ],
                  "last_updated": {
                    "block_hash": "cedd9ed0572e357af2e383382053ee40cb0ab38a07f58bae74314117c878fbd5",
                    "block_height": 5095599
                  },
                  "next_cursor": null
                }
              }
            }
          },
          "400": {
            "description": "Malformed query parameters"
          },
          "404": {
            "description": "Requested entity not found on-chain"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "DuneAndAmount": {
        "type": "object",
        "required": [
          "dune_id",
          "amount"
        ],
        "properties": {
          "amount": {
            "type": "string"
          },
          "dune_id": {
            "type": "string"
          }
        }
      },
...
      "InscriptionAndOffset": {
        "type": "object",
        "required": [
          "offset",
          "inscription_id"
        ],
        "properties": {
          "inscription_id": {
            "type": "string"
          },
          "offset": {
            "type": "integer",
            "format": "int32",
            "minimum": 0
          }
        }
      },
...
      "LastUpdated": {
        "type": "object",
        "required": [
          "block_hash",
          "block_height"
        ],
        "properties": {
          "block_hash": {
            "type": "string"
          },
          "block_height": {
            "type": "integer",
            "format": "int64",
            "minimum": 0
          }
        }
      },
...
      "PaginatedUtxo": {
        "type": "object",
        "required": [
          "data",
          "last_updated"
        ],
        "properties": {
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Utxo"
            }
          },
          "last_updated": {
            "$ref": "#/components/schemas/LastUpdated"
          },
          "next_cursor": {
            "type": "string",
            "nullable": true
          }
        }
      },
...
      "Utxo": {
        "type": "object",
        "required": [
          "txid",
          "vout",
          "script_pubkey",
          "satoshis",
          "confirmations",
          "height",
          "dunes",
          "inscriptions"
        ],
        "properties": {
          "address": {
            "type": "string",
            "nullable": true
          },
          "confirmations": {
            "type": "integer",
            "format": "int64",
            "minimum": 0
          },
          "dunes": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/DuneAndAmount"
            }
          },
          "height": {
            "type": "integer",
            "format": "int64",
            "minimum": 0
          },
          "inscriptions": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/InscriptionAndOffset"
            }
          },
          "satoshis": {
            "type": "string"
          },
          "script_pubkey": {
            "type": "string"
          },
          "txid": {
            "type": "string"
          },
          "vout": {
            "type": "integer",
            "format": "int32",
            "minimum": 0
          }
        }
      }
    }
  }
}